Added apps
