+++ frontend/src/scenes/plugins/types.ts
@@ -39,5 +39,9 @@ export enum PluginTab {
     Installed = 'installed',
     Repository = 'repository',
     Advanced = 'advanced',
+
+    // New values
+    Apps = 'apps',
+    BatchExports = 'batch_exports',
     History = 'history',
 }
